CentralApp/src/requirements.txt
web.py >= 0.38
xlrd >= 1.0.0
SQLAlchemy >= 1.1.4
Pillow >= 3.4.2
urllib3[secure]
googlemaps
openpyxl
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
